+++ arcos4py/plotting/__init__.py
@@ -4,6 +4,6 @@ __author__ = """Benjamin Graedel"""
 __email__ = "benjamin.graedel@unibe.ch"
 __version__ = '0.1.0'
 
-from .plotting import plotOriginalDetrended, dataPlots, statsPlots
+from .plotting import dataPlots, plotOriginalDetrended, statsPlots
 
 __all__ = ["plotOriginalDetrended", "dataPlots", "statsPlots"]